# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Maquettiste editor API
  version: 0.2.0-phase2
  summary: The JSON API the Maquettiste editor's C# functions expose on the editor's own domain (SPEC Section 16), phase 2 subset.
  description: |
    The editor site's `_functions/` answer this API on the editor's domain (`http://maquettiste.localhost:8080` in local
    mode). Every handler is route parsing, the sign-in check and one call into `Maquettiste.Engine`
    (`docs/engineering/engine-design.md` section 15, `docs/engineering/phase2-design.md` section 3). The binding
    design is `docs/engineering/phase2-design.md`; this file is the contract the editor SPA, its generated client and its
    MSW mocks are built from.

    **Serialization.** Every body that mirrors an engine record is that record serialized by `System.Text.Json` with
    `JsonSerializerDefaults.Web` and no custom converters (host-contracts requirement 5): camelCase names, every
    property written, `null` included, enums as the kebab-case strings the engine declares, `TimeSpan` as
    `[-][d.]hh:mm:ss[.fffffff]`, `DateTimeOffset` as ISO 8601 with offset, and non-ASCII and HTML-sensitive characters
    escaped as `\uXXXX` (the default encoder). Computed public properties are serialized too (`ChangeSet.isEmpty`,
    `ValidationReport.hasErrors`). Schemas marked `x-engine-record` name the C# record they mirror; a change to that
    record is a change to this file in the same commit.

...
    and sends `json`, never `element`.

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - editorToken: []
...
paths:
...
  /api/localization:
    get:
      tags: [localization]
      operationId: getLocalizationStatus
      summary: Localization settings and completeness per locale and shard
      description: |
        reference-types-seeds-localization.md section 3.9. The default locale, the declared locales, and per translated locale its
        fallback chain and per shard the expected, translated, missing and stale counts. A project without `localization` answers
        `defaultLocale: null` and no locales.
      x-maquettiste-role: viewer
      responses:
        '200':
          description: The settings and the completeness counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalizationStatus'
              example:
                defaultLocale: en
                locales:
                  - locale: fr
                    shards:
                      - shard: .maquettiste/model/locales/fr/_reference-data.json
                        expected: 3
                        translated: 2
                        missing: 1
                        stale: 0
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/localization/{locale}/entries:
    parameters:
      - $ref: '#/components/parameters/Locale'
    get:
      tags: [localization]
      operationId: getTranslations
      summary: Translation entries of one locale, by owner, by shard or the missing ones
      description: |
        reference-types-seeds-localization.md section 3.9. `?owner=<id>` or `?shard=<path>` returns entries with their source text,
        translation, effective text, state, shard path and hash; with neither, every entry of the locale. `?missing=true&cursor=` pages
        (200 per page) the entries that need work (`missing`, `fallback` or `stale`) for the translation queue; `cursor` is the
        previous page's last `id/field`. 404 for a locale that is not declared or is the default.
      x-maquettiste-role: viewer
      parameters:
        - name: owner
          in: query
          schema:
            $ref: '#/components/schemas/Ulid'
        - name: shard
          in: query
          schema:
            type: string
        - name: missing
          in: query
          schema:
            type: boolean
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: The entries, and the cursor of the next page when paged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationPage'
              example:
                entries:
                  - id: 01JBS3C4DWA7N36096Q14DR9GP
                    field: label
                    source: Kilogram
                    translation: Kilogramme
                    state: translated
                    shard: .maquettiste/model/locales/fr/_reference-data.json
                    shardHash: '0000000000000000000000000000000000000000000000000000000000000000'
                cursor: null
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    put:
      tags: [localization]
      operationId: putTranslations
      summary: Write, remove or confirm translations of one locale
      description: |
        reference-types-seeds-localization.md section 3.9. `{ entries: [{ id, field, value | null, confirm? }], expected: { shardPath:
        hash } }`, one atomic save: 200 with the new hashes of the shards written, 409 when a shard named in `expected` changed (its
        current hash in `hashes`, nothing written), 422 with diagnostics (MQ7202, MQ7203, MQ7211). `confirm` rewrites that field's
        `src` hash without changing the text. Each entry goes to its node's shard; an entry held in another shard moves there.
      x-maquettiste-role: editor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationWrite'
      responses:
        '200':
          description: The shards written and their new hashes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationWriteResult'
              example:
                outcome: saved
                diagnostics: []
                hashes:
                  .maquettiste/model/locales/fr/_reference-data.json: '0000000000000000000000000000000000000000000000000000000000000000'
        '409':
          description: A shard named in `expected` changed; `hashes` holds its current hash and nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationWriteResult'
        '422':
          description: The entries are invalid; `diagnostics` says why and nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationWriteResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/localization/{locale}/export:
    parameters:
      - $ref: '#/components/parameters/Locale'
    get:
      tags: [localization]
      operationId: exportTranslations
      summary: Export one locale as XLIFF 2.1 or CSV
      description: |
        reference-types-seeds-localization.md section 3.9. `?format=xliff|csv&shard=`: XLIFF 2.1 with one `file` per shard and one
        unit per field keyed `id/field` (segment `state` `translated` or `initial`, `subState` `maquettiste:<state>`), or CSV with the
        columns `id, field, source, translation, state, shard`.
      x-maquettiste-role: viewer
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [xliff, csv]
            default: xliff
        - name: shard
          in: query
          schema:
            type: string
      responses:
        '200':
          description: The exported file.
          content:
            application/xliff+xml:
              schema:
                type: string
            text/csv:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/localization/{locale}/import:
    parameters:
      - $ref: '#/components/parameters/Locale'
    post:
      tags: [localization]
      operationId: importTranslations
      summary: Preview or apply an XLIFF or CSV translation import
      description: |
        reference-types-seeds-localization.md section 3.9. `?dryRun=true` (the default) previews; `dryRun=false` writes the
        translations as one save checked against the shard hashes read for the preview (409 when one changed). A unit with an empty
        target is left alone; a unit naming no translatable field is ignored with MQ7203 in `diagnostics`.
      x-maquettiste-role: editor
      parameters:
        - name: dryRun
          in: query
          schema:
            type: boolean
            default: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format, content]
              properties:
                format:
                  type: string
                  enum: [xliff, csv]
                content:
                  type: string
      responses:
        '200':
          description: The preview (or the result) of the import.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
              example:
                added: 1
                changed: []
                removed: 0
                diagnostics: []
        '409':
          description: What the preview read changed before the import was applied; nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '422':
          description: The import makes the model invalid; `diagnostics` says why and nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/seeds/{id}/csv:
    parameters:
      - $ref: '#/components/parameters/ElementId'
    get:
      tags: [reference-data]
      operationId: exportSeedCsv
      summary: Export a seed's rows as CSV
      description: |
        reference-types-seeds-localization.md sections 2.3 and 3.9. RFC 4180, UTF-8, `@id` first, then `@code`, `@label` and `@description` for reference types, then attribute and end
        role names in the seed's column order, then `@label:<locale>` and `@description:<locale>` per `locale`; `?bom=true` adds a
        byte order mark and CRLF line ends. Reference cells export codes, collections `;`-joined codes, end cells row ids, value
        objects compact JSON.
      x-maquettiste-role: viewer
      parameters:
        - name: bom
          in: query
          schema:
            type: boolean
            default: false
        - name: locale
          in: query
          description: Adds `@label:<locale>` and `@description:<locale>` columns; repeatable.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: The CSV text.
          content:
            text/csv:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    post:
      tags: [reference-data]
      operationId: importSeedCsv
      summary: Preview or apply a CSV import into a seed
      description: |
        reference-types-seeds-localization.md sections 2.3 and 3.9. The body is JSON (`{ content }`, the CSV text), as every write
        of the API. Rows match by `@id`, else by `@code` for reference types, else are new rows with new ids; `merge` updates and
        adds, `replace` also removes rows missing from the file, except rows that cells of other seeds name (listed in `blocked`).
        Headers that match no column are listed in `ignoredHeaders`. `?dryRun=true` (the default) previews; applying saves the seed
        with `If-Match` (or the hash read for the preview; 409 when it changed, 422 with diagnostics) and then writes the
        `@label:<locale>` and `@description:<locale>` translations.
      x-maquettiste-role: editor
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [merge, replace]
            default: merge
        - name: dryRun
          in: query
          schema:
            type: boolean
            default: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: The CSV text.
      responses:
        '200':
          description: The preview (or the result) of the import.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
              example:
                added: 1
                changed: []
                removed: 0
                diagnostics: []
        '409':
          description: What the preview read changed before the import was applied; nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '422':
          description: The import makes the model invalid; `diagnostics` says why and nothing was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/reference-types/{id}/usage:
    parameters:
      - $ref: '#/components/parameters/ElementId'
    get:
      tags: [reference-data]
      operationId: getReferenceTypeUsage
      summary: The attributes that use a reference type, with their effective storage
      description: |
        reference-types-seeds-localization.md section 3.9. Every attribute typed by the reference type, with its owner, domain, collection, required and the effective storage
        choice per database.
      x-maquettiste-role: viewer
      responses:
        '200':
          description: The usages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceTypeUsage'
              example:
                usages: []
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mq_session
      description: Set by `POST /api/session`. Host-only, HttpOnly, SameSite=Strict; protected with the host's data-protection keys; carries the user name, role, issue time and a fingerprint of the editor token.
    editorToken:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <MAQUETTISTE_EDITOR_TOKEN>`, the secret site variable. For the CLI, scripts and tests. Compared in fixed time. In local mode a loopback or `MAQUETTISTE_LOCAL_PEERS` caller with a local `Host` and no forwarding header needs neither scheme and is the local admin (see the authentication notes in `info`).'
  parameters:
    Locale:
      name: locale
      in: path
      required: true
      description: A declared BCP 47 locale other than the default.
      schema:
        type: string
        pattern: '^[A-Za-z]{2,8}(-[A-Za-z0-9]{1,8})*$'
    ElementId:
      name: id
      in: path
      required: true
      description: An element or sub-element id (uppercase ULID).
      schema:
        $ref: '#/components/schemas/Ulid'
      example: 01J92P0V0FJ23CGSNKM7P1W5V7
...
  responses:
    BadRequest:
      description: The body is not JSON or not the expected shape (`bad-request`), or is larger than 4 MB (`too-large`). The sign-in gate refuses a `POST` or `PUT` that is not `application/json` with 415 `unsupported-media-type`, and a foreign `Origin` with 403 `forbidden-origin`, before any handler runs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: The request body is not valid JSON.
            status: 400
            code: bad-request
            detail: "'}' is invalid after a value. Expected either ',', '}', or ']'. LineNumber: 3 | BytePositionInLine: 2."
    Unauthenticated:
      description: No valid credential (`unauthenticated`), or a wrong token (`bad-token`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Sign in to use the editor.
            status: 401
            code: unauthenticated
    NotFound:
      description: No such element, diagram, plan, path in the plan, job or connection (`not-found`, `not-a-diagram`, `not-a-database`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: No element has the id 01J92P0V0FJ23CGSNKM7P1W5V9.
            status: 404
            code: not-found
...
  schemas:
...
    Ulid:
      description: Uppercase Crockford ULID (engine-design.md D1).
      type: string
      pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
...
    LocalizationStatus:
      description: Localization settings and completeness (reference-types-seeds-localization.md section 3.9).
      type: object
      required: [defaultLocale, locales]
      properties:
        defaultLocale:
          type: [string, 'null']
        declared:
          type: array
          description: Every declared locale, the default included, as the settings list them.
          items:
            type: string
        locales:
          type: array
          description: The locales other than the default.
          items:
            type: object
            required: [locale, shards]
            properties:
              locale:
                type: string
              chain:
                type: array
                description: The fallback chain, the locale first and the default last.
                items:
                  type: string
              shards:
                type: array
                items:
                  type: object
                  required: [shard, expected, translated, missing, stale]
                  properties:
                    shard:
                      type: string
                    expected:
                      type: integer
                    translated:
                      type: integer
                    missing:
                      type: integer
                    stale:
                      type: integer
    TranslationPage:
      type: object
      required: [entries, cursor]
      properties:
        entries:
          type: array
          items:
            type: object
            required: [id, field, source, translation, state, shard, shardHash]
            properties:
              id:
                $ref: '#/components/schemas/Ulid'
              owner:
                description: The element whose file holds the node.
                $ref: '#/components/schemas/Ulid'
              effective:
                type: [string, 'null']
                description: The text the fallback chain gives.
              field:
                type: string
                enum: [displayName, pluralName, label, description]
              source:
                type: [string, 'null']
              translation:
                type: [string, 'null']
              state:
                type: string
                enum: [translated, missing, stale, fallback]
              shard:
                type: string
              shardHash:
                oneOf:
                  - $ref: '#/components/schemas/Hash'
                  - type: 'null'
        cursor:
          type: [string, 'null']
    TranslationWrite:
      type: object
      required: [entries, expected]
      properties:
        entries:
          type: array
          items:
            type: object
            required: [id, field, value]
            properties:
              id:
                $ref: '#/components/schemas/Ulid'
              field:
                type: string
                enum: [displayName, pluralName, label, description]
              value:
                description: 'The text, `{ "file": "<sidecar>" }`, or `null` to remove.'
              confirm:
                type: boolean
        expected:
          type: object
          description: Shard path to the hash the caller loaded.
          additionalProperties:
            $ref: '#/components/schemas/Hash'
    TranslationWriteResult:
      description: The answer of a translation write.
      type: object
      required: [outcome, hashes, diagnostics]
      properties:
        outcome:
          type: string
          enum: [saved, conflict, invalid]
        hashes:
          type: object
          description: The new hashes of the shards written (saved), or the current hashes of the shards that changed (conflict).
          additionalProperties:
            $ref: '#/components/schemas/Hash'
        diagnostics:
          type: array
          items:
            $ref: '#/components/schemas/Diagnostic'
...